ci: gate the eight adapter crates, and make three checks real
Four gaps, each of which let a check exist without protecting anything.

1. Eight crates had no clippy gate. `linux.yml` lists five frontend crates;
   the Android and browser adapters opt into `[lints] workspace = true` like
   the rest but were never added when they were introduced. They are clean
   today (two real findings, fixed below), so they go in a SEPARATE step: the
   existing five carry ~43 pre-existing findings in neomacs-layout-engine and
   six in neomacs-display-protocol, and merging the lists would have made the
   new coverage hostage to unrelated debt. The display-protocol six are fixed
   here since they blocked verifying anything; the layout-engine ones are left
   alone, being neither this branch's nor a one-line matter.

2. The three browser regression gates were unpinned. `main_test.rs` checks the
   shape of the `wasm-browser-smoke` job but never mentioned
   browser_stack_smoke.py, browser_opfs_write_smoke.py or
   browser_input_rejection_smoke.py -- so deleting the steps that guard the
   deep-recursion trap, the file-zeroing save and the poison-batch hot-spin
   still passed. Now pinned, with browser_hidpi_smoke.py and the --call-style
   sweep.

3. The browser HTTP ERT never ran. It sat in `file-notify-public-ert`, whose
   matrix is macOS + Windows only, behind a 180-minute fresh build, so a
   Linux-only run executed it zero times. Moved to `neomacs-test-runtime`,
   which already builds a release runtime and runs ERT suites. Verified
   locally against target/release/neomacs: 4 tests, 4 as expected. This is
   the first time they have actually run.

4. ViewportProbe was never even compiled. It is the only runtime check of the
   packaged Android app, wired as `testInstrumentationRunner`, but
   `assembleRelease` does not build `androidTest` -- so it could have rotted
   silently. The android-package job now also runs
   `assembleDebugAndroidTest`, and that is pinned too. Executing it still
   needs an emulator on the runner, which is a separate decision.

Clippy fixes: an `unwrap_or_else` with a side-effect-free fallback, four
`then(|| cheap).flatten()` -> `then_some`, an item-scoped allow for
`measure_panel`'s eight arguments (moved verbatim from the desktop menu code;
reshaping the signature here would diverge it from the original), and an
allow for `drop(push)` in glyph_matrix, where dropping the closure is what
ends its mutable borrow -- a borrow release clippy's `drop_non_drop` cannot
see, so removing it would not compile.

neovm-core 10075/10075, adapter+frontend crates 4904/4904, JS 50/50, the new
clippy step exits 0. The five `ci_*` shape guards in xtask remain red exactly
as they are on origin/main -- same assertion messages -- and are not touched
here; they assert strings the workflow has not contained for some time
(`/usr/bin/emacs`, `CARGO_BUILD_JOBS: "1"` in the archive job) and only their
author can say which side is wrong.

+++ crates/neomacs-wgpu-runtime/src/surface_frame.rs
@@ -255,7 +255,7 @@ impl SurfaceFrameRenderer {
                 let cycle = &frame
                     .effective_window_cursor_effects(cursor.window_id, &self.renderer.effects)
                     .cursor_color_cycle;
-                cycle.enabled.then_some(u32::from(cycle.fps.get()).min(60))
+                cycle.enabled.then_some(cycle.fps.get().min(60))
             })
             .max();
         motion_rate
